Cast arguments to setPriority() to quiet compiler.

+++ TreePiece.cpp
@@ -1829,7 +1829,7 @@ void TreePiece::startORBTreeBuild(CkReductionMsg* m){
     if (node->getType() == Empty || node->moments.totalMass > 0) {
       for (int i=0; i<l->length(); ++i) {
           CkEntryOptions opts;
-          opts.setPriority(-100000000);
+          opts.setPriority((unsigned int) -100000000);
 	  streamingProxy[(*l)[i]].receiveRemoteMoments(nodeKey, node->getType(), node->firstParticle, node->particleCount, node->moments, node->boundingBox, node->bndBoxBall, node->iParticleTypes, &opts);
 	//CkPrintf("[%d] sending moments of %s to %d upon treebuild finished\n",thisIndex,keyBits(node->getKey(),63).c_str(),(*l)[i]);
       }
@@ -1929,7 +1929,7 @@ void TreePiece::buildORBTree(GenericTreeNode * node, int level){
 	      node->remoteIndex --;
 	      // request the remote chare to fill this node with the Moments
               CkEntryOptions opts;
-              opts.setPriority(-110000000);
+              opts.setPriority((unsigned int) -110000000);
 	      streamingProxy[child->remoteIndex].requestRemoteMoments(child->getKey(), thisIndex, &opts);
 	      //CkPrintf("[%d] asking for moments of %s to %d\n",thisIndex,keyBits(child->getKey(),63).c_str(),child->remoteIndex);
       }
@@ -2157,7 +2157,7 @@ void TreePiece::sendRequestForNonLocalMoments(GenericTreeNode *pickedNode){
     pickedNode->remoteIndex = getResponsibleIndex(first,last);
     // request the remote chare to fill this node with the Moments
     CkEntryOptions opts;
-    opts.setPriority(-110000000);
+    opts.setPriority((unsigned int) -110000000);
     //CkPrintf("[%d] phase I request moments for %llu from %d\n", tp->thisIndex, node->getKey(), node->remoteIndex);
     treeProxy[pickedNode->remoteIndex].requestRemoteMoments(pickedNode->getKey(), thisIndex, &opts);
   }
@@ -2206,7 +2206,7 @@ void TreePiece::processRemoteRequestsForMoments(){
     if (node->getType() == Empty || node->moments.totalMass > 0) {
       for (int i=0; i<l->length(); ++i) {
           CkEntryOptions opts;
-          opts.setPriority(-100000000);
+          opts.setPriority((unsigned int) -100000000);
 	  streamingProxy[(*l)[i]].receiveRemoteMoments(nodeKey, node->getType(), node->firstParticle, node->particleCount, node->moments, node->boundingBox, node->bndBoxBall, node->iParticleTypes, &opts);
 	  //CkPrintf("[%d] sending moments of %s to %d upon treebuild finished\n",thisIndex,keyBits(node->getKey(),63).c_str(),(*l)[i]);
       }
@@ -2404,7 +2404,7 @@ void TreePiece::buildOctTree(GenericTreeNode * node, int level) {
 	node->remoteIndex --;
 	// request the remote chare to fill this node with the Moments
         CkEntryOptions opts;
-        opts.setPriority(-110000000);
+        opts.setPriority((unsigned int) -110000000);
 	streamingProxy[child->remoteIndex].requestRemoteMoments(child->getKey(), thisIndex, &opts);
 	//CkPrintf("[%d] asking for moments of %s to %d\n",thisIndex,keyBits(child->getKey(),63).c_str(),child->remoteIndex);
       }
@@ -2509,7 +2509,7 @@ void TreePiece::requestRemoteMoments(const Tree::NodeKey key, int sender) {
   GenericTreeNode *node = keyToNode(key);
   if (node != NULL && (node->getType() == Empty || node->moments.totalMass > 0)) {
       CkEntryOptions opts;
-      opts.setPriority(-100000000);
+      opts.setPriority((unsigned int) -100000000);
       streamingProxy[sender].receiveRemoteMoments(key, node->getType(), node->firstParticle, node->particleCount, node->moments, node->boundingBox, node->bndBoxBall, node->iParticleTypes, &opts);
     //CkPrintf("[%d] sending moments of %s to %d directly\n",thisIndex,keyBits(node->getKey(),63).c_str(),sender);
   } else {
@@ -2612,7 +2612,7 @@ GenericTreeNode *TreePiece::boundaryParentReady(GenericTreeNode *parent){
     CkVec<int> *l = iter->second;
     for (int i=0; i<l->length(); ++i) {
       CkEntryOptions opts;
-      opts.setPriority(-100000000);
+      opts.setPriority((unsigned int) -100000000);
       streamingProxy[(*l)[i]].receiveRemoteMoments(parent->getKey(), parent->getType(), parent->firstParticle, parent->particleCount, parent->moments, parent->boundingBox, parent->bndBoxBall, parent->iParticleTypes, &opts);
       //CkPrintf("[%d] sending moments of %s to %d\n",thisIndex,keyBits(parent->getKey(),63).c_str(),(*l)[i]);
     }
